pyproject.toml
[project]
name = "oobot"
version = "0.1.0"
description = "Telegram Bot for monitoring OpenCode sessions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "libtmux>=0.37.0",
    "Pillow>=10.0.0",
    "telegramify-markdown>=0.5.0",
    "aiofiles>=24.0.0",
]

[project.scripts]
oobot = "oobot.main:main"

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oobot"]
